pyproject.toml
[tool.poetry]
name = "amazing-hunting"
version = "0.1.0"
description = ""
authors = ["alexeydoronin <neonixys@mail.ru>"]
readme = "README.md"
packages = [{include = "amazing_hunting"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
serializers = "^0.2.4"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
psycopg2 = "^2.9.5"
django = "^4.1.6"
pytest-factoryboy = "^2.5.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
